Added a cutoff for the I band magnitude, temporarily set to 24.5.
Note: without a cutoff, the simulation required 4 days of computing time on the irsagpu1 machine.
Approximately 1 day was required to generate the model grid. If the model grid is pre-computed and
a magnitude cutoff of ~24.5 is used, then the simulation will still require approximately 1 day of
computing time on irsagpu1, unless the grid is further optimized. Using more machines improves the
situation. However, the only way to do this currently, is by using a StratOS / NebulOS cluster or
doing the job manually. It should be fairly straightforward to distribute the work using a Slurm
cluster, if one becomes available.